% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_canada_2022_electoral_ridings.R
\docType{data}
\name{census_canada_2022_electoral_ridings}
\alias{census_canada_2022_electoral_ridings}
\title{Données de recensement structurées – Circonscriptions électorales fédérales (Canada, 2022)}
\format{
Un data.frame avec 9 261 lignes et 5 colonnes :
\describe{
\item{id_riding}{Identifiant numérique de la circonscription électorale fédérale}
\item{variable}{Nom de la variable sociodémographique (ex: \code{ses_age}, \code{ses_gender}, \code{ses_income})}
\item{category}{Catégorie de la variable (ex: \verb{18_24}, \code{female}, \verb{30000_to_50000})}
\item{count}{Nombre de personnes dans cette catégorie}
\item{prop}{Proportion de la population de la circonscription dans cette catégorie (entre 0 et 1)}
}
}
\source{
Statistique Canada
}
\usage{
census_canada_2022_electoral_ridings
}
\description{
Données sociodémographiques issues du recensement canadien de 2021, structurées par circonscription électorale fédérale
selon le redécoupage de 2022. Chaque ligne correspond à une catégorie d'une variable dans une circonscription donnée.
}
\keyword{datasets}
