% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_canada_2022_electoral_rta.R
\docType{data}
\name{census_canada_2022_rta}
\alias{census_canada_2022_rta}
\title{Données de recensement structurées – RTA (Canada, 2021)}
\format{
Un data.frame avec 9 261 lignes et 5 colonnes :
\describe{
\item{rta}{Identifiant de la RTA}
\item{variable}{Nom de la variable sociodémographique (ex: \code{ses_age}, \code{ses_gender}, \code{ses_income})}
\item{category}{Catégorie de la variable (ex: \verb{18_24}, \code{female}, \verb{30000_to_50000})}
\item{count}{Nombre de personnes dans cette catégorie}
\item{prop}{Proportion de la population de la RTA dans cette catégorie (entre 0 et 1)}
}
}
\source{
Statistique Canada
}
\usage{
census_canada_2022_rta
}
\description{
Données sociodémographiques issues du recensement canadien de 2021, structurées par région de tri d'acheminement (RTA).
Chaque ligne correspond à une catégorie d'une variable dans une RTA donnée.
}
\keyword{datasets}
