% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{intersect_spatial_objects}
\alias{intersect_spatial_objects}
\title{Compute Spatial Intersections and Coverage Proportions

This function computes the intersection between two spatial objects and calculates the area covered
by each unit of \code{spatial_target} within each unit of \code{spatial_ref}.}
\usage{
intersect_spatial_objects(
  spatial_ref,
  id_ref,
  spatial_target,
  id_target,
  dTolerance = 50
)
}
\arguments{
\item{spatial_ref}{An \code{sf} object representing the reference spatial units that will be intersected.}

\item{id_ref}{A string specifying the column name in \code{spatial_ref} that contains unique identifiers.}

\item{spatial_target}{An \code{sf} object representing the target spatial units used to intersect \code{spatial_ref}.}

\item{id_target}{A string specifying the column name in \code{spatial_target} that contains unique identifiers.}

\item{dTolerance}{A numeric value specisi j'utfying the simplification tolerance for spatial geometries before computing intersections. Default is \code{50}.}
}
\value{
A tibble containing the intersections with the following columns:
\itemize{
\item \code{{id_ref}}: The unique identifier of the reference spatial unit.
\item \code{{id_target}}: The unique identifier of the intersecting target spatial unit.
\item \code{area_covered_by_target_m2}: The intersection area in square meters.
\item \code{prop_of_ref_area_covered_by_target}: The proportion of the reference unit covered by the target unit.
}
}
\description{
Compute Spatial Intersections and Coverage Proportions

This function computes the intersection between two spatial objects and calculates the area covered
by each unit of \code{spatial_target} within each unit of \code{spatial_ref}.
}
\details{
\itemize{
\item The function simplifies the geometries before computing intersections to improve performance.
\item It issues warnings if no intersections are found or if some reference units are missing in the final result.
\item The user can adjust \code{dTolerance} to control the level of simplification.
}
}
\examples{
 
\dontrun{  
 
# Example spatial objects (must be valid sf objects)  
result <- intersect_spatial_objects(spatial_ref, "region_id", spatial_target, "district_id")  
}  
 
}
