% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_spatial_target_from_origin.R
\name{prepare_simulations}
\alias{prepare_simulations}
\title{Simuler des répondants selon les distributions du recensement et la géographie intersectionnelle}
\usage{
prepare_simulations(
  census_origin,
  census_target,
  spatial_intersection,
  origin_col,
  target_col,
  n_sim = 500,
  origins = NULL
)
}
\arguments{
\item{census_origin}{Données de recensement agrégées par unité spatiale d'origine, au format long
(\code{variable}, \code{category}, \code{count}, \code{prop}).}

\item{census_target}{Données de recensement agrégées par unité spatiale cible, au même format que \code{census_origin}.}

\item{spatial_intersection}{Table des intersections spatiales produite par
\code{cartessn::intersect_spatial_objects()}.}

\item{origin_col}{Nom de la colonne identifiant l’unité spatiale d’origine (ex. \code{"rta"}).}

\item{target_col}{Nom de la colonne identifiant l’unité spatiale cible (ex. \code{"id_riding"}).}

\item{n_sim}{Nombre de répondants à simuler par unité spatiale d'origine.}

\item{origins}{(Optionnel) Vecteur de noms/id des unités spatiales d'origine à traiter. Si \code{NULL}, toutes sont utilisées.}
}
\value{
Une liste contenant :
\describe{
\item{cases}{Vecteur nommé identifiant les cas : 1 (une seule cible), 2 (plusieurs).}
\item{unique_predictions}{Liste nommée de vecteurs de probabilité (valeur 1) pour les cas à une seule cible.}
\item{simulated_datasets}{Liste nommée de jeux de données simulés pour les cas à plusieurs cibles.}
}
}
\description{
Cette fonction génère des données simulées ou des prédictions uniques par unité spatiale d'origine (\code{origin_col}),
selon le nombre de cibles (\code{target_col}) qui lui sont associées. Elle retourne un objet contenant les cas,
les prédictions uniques (si une seule cible), et les jeux de données simulés (si plusieurs).
}
