% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_spatial_target_from_origin.R
\name{simulate_respondents_from_census_unit}
\alias{simulate_respondents_from_census_unit}
\title{Simuler des répondants à partir de données de recensement spatiales}
\usage{
simulate_respondents_from_census_unit(
  origin_unit_id,
  n = 500,
  spatial_intersection,
  origin_id_col = "origin",
  target_id_col = "target",
  census_origin,
  census_target
)
}
\arguments{
\item{origin_unit_id}{Identifiant d'une unité spatiale d'origine (ex. : "G2L" pour une RTA).}

\item{n}{Nombre de répondants à simuler (par défaut : 500).}

\item{spatial_intersection}{Data frame indiquant la proportion de chaque unité cible couverte
par chaque unité d'origine. Doit contenir les colonnes \code{origin_id_col}, \code{target_id_col},
et \code{prop_of_ref_area_covered_by_target}.}

\item{origin_id_col}{Nom de la colonne contenant l'identifiant de l'unité spatiale d'origine (ex. : \code{"rta"}).}

\item{target_id_col}{Nom de la colonne contenant l'identifiant de l'unité spatiale cible (ex. : \code{"id_riding"}).}

\item{census_origin}{Données de recensement associées aux unités d'origine. Doit contenir les colonnes :
\code{variable}, \code{category}, \code{prop}, ainsi qu’une colonne correspondant à \code{origin_id_col}.}

\item{census_target}{Données de recensement associées aux unités cibles. Doit contenir les colonnes :
\code{variable}, \code{category}, \code{prop}, ainsi qu’une colonne correspondant à \code{target_id_col}.}
}
\value{
Un \code{tibble} contenant \code{n} répondants simulés avec :
\itemize{
\item un identifiant unique (\code{id}),
\item l’unité spatiale d’origine (\code{origin_id_col}),
\item l’unité spatiale cible assignée (\code{target_id_col}),
\item et les variables sociodémographiques simulées selon les distributions croisées.
}
}
\description{
Cette fonction génère un ensemble de répondants simulés à partir des distributions
sociodémographiques disponibles pour une unité spatiale d'origine (\code{origin_unit_id})
en combinant les données de recensement de cette unité et des unités cibles qu'elle recouvre,
pondérées selon la superficie partagée.
}
\examples{
\dontrun{
simulate_respondents_from_census_unit(
  origin_unit_id = "G2L",
  n = 500,
  spatial_intersection = df_intersections,
  origin_id_col = "rta",
  target_id_col = "id_riding",
  census_origin = df_census_rta,
  census_target = df_census_ridings
)
}

}
