% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_canada_2013_electoral_ridings.R
\docType{data}
\name{spatial_canada_2013_electoral_ridings}
\alias{spatial_canada_2013_electoral_ridings}
\title{Données géospatiales du Canada – Circonscriptions électorales (redécoupage de 2013)}
\format{
Un objet \code{sf} avec une colonne \code{geometry} représentant les polygones des circonscriptions.
}
\source{
Statistique Canada
}
\usage{
spatial_canada_2013_electoral_ridings
}
\description{
Objet sf contenant les circonscriptions électorales du Canada, fusionnées à partir de plusieurs sources.
}
\keyword{datasets}
