% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_canada_2021_rta.R
\docType{data}
\name{spatial_canada_2021_rta}
\alias{spatial_canada_2021_rta}
\title{Données géospatiales du Canada – régions de tri d'acheminement (RTA)}
\format{
Un objet \code{sf} avec une colonne \code{geometry} représentant les polygones des RTA
}
\source{
Statistique Canada
}
\usage{
spatial_canada_2021_rta
}
\description{
Objet sf contenant les régions de tri d'acheminement du Canada (RTA).
}
\keyword{datasets}
